# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MykytaDu API
  version: 0.3.1
  description: |
    Contrato HTTP inicial do MykytaDu API.

    Esta versão descreve as superfícies iniciais de Identity e Translation.
    A API usa JSON em camelCase, IDs UUID como strings e datas ISO-8601 em
    UTC. O prefixo `/api/v1` faz parte da URL base do servidor.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8081/api/v1
    description: Desenvolvimento local
...
paths:
  /auth/register:
    post:
      tags: [Identity]
      operationId: registerAccount
      summary: Criar conta
      description: |
        Cria uma conta no estado `pending` e inicia a verificação de e-mail.
        O endpoint não retorna credenciais nem tokens. A resposta é uniforme
        para não transformar o texto de erro em mecanismo de enumeração.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              default:
                $ref: '#/components/examples/RegisterRequest'
      responses:
        '201':
          description: Conta criada e aguardando verificação de e-mail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
              examples:
                default:
                  $ref: '#/components/examples/RegistrationResponse'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '409':
          description: Cadastro não pode ser concluído
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                registrationRejected:
                  $ref: '#/components/examples/RegistrationRejectedProblem'
        '429':
          $ref: '#/components/responses/RateLimitProblem'
        '503':
          $ref: '#/components/responses/EmailDeliveryProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /auth/verify-email:
    post:
      tags: [Identity]
      operationId: verifyEmail
      summary: Verificar e-mail
      description: Consome um token de ação válido e ativa a conta atomicamente.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionTokenRequest'
            examples:
              default:
                $ref: '#/components/examples/ActionTokenRequest'
      responses:
        '204':
          description: E-mail verificado e conta ativada
        '400':
          description: Token ausente, inválido, expirado ou já consumido
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                invalidActionToken:
                  $ref: '#/components/examples/InvalidActionTokenProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /auth/verify-email/resend:
    post:
      tags: [Identity]
      operationId: resendEmailVerification
      summary: Reemitir verificação de e-mail
      description: |
        Aceita a solicitação de forma uniforme, exista ou não uma conta
        `pending` elegível. Para uma conta elegível, invalida tokens anteriores,
        cria um novo token e tenta o envio somente após o commit. O endpoint não
        revela existência, estado da conta nem resultado do provedor de e-mail.
      security: []
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailVerificationResendRequest'
            examples:
              default:
                $ref: '#/components/examples/EmailVerificationResendRequest'
      responses:
        '202':
          description: Solicitação aceita independentemente da elegibilidade da conta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionAcceptedResponse'
              examples:
                default:
                  $ref: '#/components/examples/ActionAcceptedResponse'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '429':
          $ref: '#/components/responses/RateLimitProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT curto, mantido somente em memória no cliente.
    refreshCookie:
      type: apiKey
      in: cookie
      name: mykytadu_refresh
      description: Refresh token opaco HttpOnly, Secure, SameSite=Lax e Path=/api/v1/auth.
  parameters:
...
    Origin:
      name: Origin
      in: header
      required: false
      schema:
        type: string
        format: uri
      description: Origem enviada pelo navegador e validada pela allowlist do ambiente.
...
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Endereço normalizado pelo servidor antes da persistência.
          example: pessoa@example.com
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128
          writeOnly: true
          description: Senha recebida somente para derivação do hash Argon2id.
        displayName:
          type: [string, 'null']
          minLength: 1
          maxLength: 80
          description: Nome exibido no perfil; pode ser omitido no cadastro.
          example: Pessoa MykytaDu
    RegistrationResponse:
      type: object
      additionalProperties: false
      required: [user, nextAction]
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        nextAction:
          type: string
          enum: [verifyEmail]
          description: Próxima ação esperada do cliente.
    EmailVerificationResendRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Endereço usado somente para localizar uma conta elegível; a resposta é sempre uniforme.
          example: pessoa@example.com
...
    ActionTokenRequest:
      type: object
      additionalProperties: false
      required: [actionToken]
      properties:
        actionToken:
          type: string
          minLength: 1
          maxLength: 512
          writeOnly: true
          description: Token puro recebido por canal de ação; nunca é persistido em texto puro.
          example: <action-token-from-email>
...
    UserProfile:
      type: object
      additionalProperties: false
      required: [id, email, displayName, status, roles, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: 01991f18-7d42-7b21-a2ef-1d8e6e14a901
        email:
          type: string
          format: email
          readOnly: true
          example: pessoa@example.com
        displayName:
          type: [string, 'null']
          readOnly: true
          example: Pessoa MykytaDu
        status:
          $ref: '#/components/schemas/AccountStatus'
        roles:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Role'
        emailVerifiedAt:
          type: [string, 'null']
          format: date-time
          readOnly: true
          description: Instante UTC da verificação; nulo enquanto a conta estiver `pending`.
          example: 2026-09-17T14:30:00Z
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: 2026-09-17T14:00:00Z
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: 2026-09-17T14:30:00Z
    AccountStatus:
      type: string
      enum: [pending, active, blocked, deleted]
    Role:
      type: string
      enum: [USER, ADMIN]
...
    ActionAcceptedResponse:
      type: object
      additionalProperties: false
      required: [accepted]
      properties:
        accepted:
          type: boolean
          example: true
    ProblemCode:
      type: string
      description: Código estável usado pelo cliente; nunca substituir por `detail`, `title` ou `errors[].message`.
      enum:
        - request_validation_failed
        - registration_rejected
        - invalid_action_token
        - email_delivery_unavailable
        - authentication_required
        - invalid_credentials
        - email_verification_required
        - session_invalid
        - csrf_invalid
        - authorization_denied
        - rate_limit_exceeded
        - translation_quota_exceeded
        - translation_provider_failed
        - translation_provider_timeout
        - internal_error
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, traceId, errors]
      properties:
        type:
          type: string
          format: uri-reference
          description: URN estável da categoria do problema.
          example: urn:mykytadu:problem:request_validation_failed
        title:
          type: string
          description: Título humano curto; não usar como chave de decisão.
          example: Request validation failed
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
          example: 400
        detail:
          type: string
          description: Mensagem humana que pode variar; não usar como chave de decisão.
          example: One or more fields are invalid.
        instance:
          type: string
          description: Path da requisição sem query string.
          example: /api/v1/auth/register
        code:
          $ref: '#/components/schemas/ProblemCode'
          description: Código estável usado pelo cliente.
          example: request_validation_failed
        traceId:
          type: string
          format: uuid
          description: Identificador de correlação para suporte.
          example: 2d46eea5-6ff7-4ee8-95c9-5c71c9f82484
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
          example: password
        code:
          type: string
          example: too_short
        message:
          type: string
          description: Mensagem humana de validação; não usar como chave de decisão.
          example: must contain at least 12 characters
...
  responses:
    ValidationProblem:
      description: Corpo ausente, inválido ou com campos que não atendem às restrições
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            default:
              $ref: '#/components/examples/ValidationProblem'
...
    RateLimitProblem:
      description: Limite operacional excedido
      headers:
        Retry-After:
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Segundos mínimos antes de uma nova tentativa.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            default:
              $ref: '#/components/examples/RateLimitProblem'
    EmailDeliveryProblem:
      description: Conta persistida, mas a primeira entrega de verificação não foi aceita pelo provedor
      headers:
        Retry-After:
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Segundos recomendados antes de solicitar reemissão.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            default:
              $ref: '#/components/examples/EmailDeliveryProblem'
    InternalProblem:
      description: Falha inesperada sem detalhes internos
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            default:
              $ref: '#/components/examples/InternalProblem'
...
  examples:
    RegisterRequest:
      value:
        email: pessoa@example.com
        password: senha-fornecida-pelo-usuario
        displayName: Pessoa MykytaDu
    RegistrationResponse:
      value:
        user:
          id: 01991f18-7d42-7b21-a2ef-1d8e6e14a901
          email: pessoa@example.com
          displayName: Pessoa MykytaDu
          status: pending
          roles: [USER]
          emailVerifiedAt: null
          createdAt: 2026-09-17T14:00:00Z
          updatedAt: 2026-09-17T14:00:00Z
        nextAction: verifyEmail
    ActionTokenRequest:
      value:
        actionToken: <action-token-from-email>
    EmailVerificationResendRequest:
      value:
        email: pessoa@example.com
    ActionAcceptedResponse:
      value:
        accepted: true
...
    ValidationProblem:
      value:
        type: urn:mykytadu:problem:request_validation_failed
        title: Request validation failed
        status: 400
        detail: One or more fields are invalid.
        instance: /api/v1/auth/register
        code: request_validation_failed
        traceId: 2d46eea5-6ff7-4ee8-95c9-5c71c9f82484
        errors:
          - field: password
            code: too_short
            message: must contain at least 12 characters
...
    RegistrationRejectedProblem:
      value:
        type: urn:mykytadu:problem:registration_rejected
        title: Registration unavailable
        status: 409
        detail: The account cannot be created with the supplied data at this time.
        instance: /api/v1/auth/register
        code: registration_rejected
        traceId: 71c0e13d-a3d8-45c3-8f68-e5a7e04b5bb1
        errors: []
    InvalidActionTokenProblem:
      value:
        type: urn:mykytadu:problem:invalid_action_token
        title: Action token invalid
        status: 400
        detail: The action token is invalid or expired.
        instance: /api/v1/auth/verify-email
        code: invalid_action_token
        traceId: 88fb0ef8-9c8b-4b18-b77a-bd7e38cd1a04
        errors: []
    EmailDeliveryProblem:
      value:
        type: urn:mykytadu:problem:email_delivery_unavailable
        title: Email delivery unavailable
        status: 503
        detail: The account was created, but the verification email could not be delivered now.
        instance: /api/v1/auth/register
        code: email_delivery_unavailable
        traceId: 0f11495a-a1af-4d29-8450-e1af5d403616
        errors: []
...
    RateLimitProblem:
      value:
        type: urn:mykytadu:problem:rate_limit_exceeded
        title: Too many requests
        status: 429
        detail: The operation is temporarily rate limited.
        instance: /api/v1/auth/login
        code: rate_limit_exceeded
        traceId: 1a314b4a-1026-4c44-a4ea-4ad8c2e7b69c
        errors: []
    InternalProblem:
      value:
        type: urn:mykytadu:problem:internal_error
        title: Internal error
        status: 500
        detail: The request could not be completed.
        instance: /api/v1/auth/login
        code: internal_error
        traceId: 6d72c7b5-95f8-49b6-8e7a-0f7c6de4b2fa
        errors: []
...